Feat: Refactor API fetching logic and implement centralized error handling
