+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import { Toaster } from "@/components/ui/sonner";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import { SWRConfig } from "swr";
-import { useSWRFetcher } from "@/lib/hooks/useSWRFetcher";
+import { swrFetcher } from "@/lib/swrFetcher";
 import "./globals.css";
 
 const geistSans = Geist({
@@ -25,14 +25,12 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const fetcher = useSWRFetcher();
-
   return (
     <html lang="en">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        <SWRConfig value={{ fetcher }}>
+        <SWRConfig value={{ fetcher: swrFetcher }}>
           {children}
         </SWRConfig>
         <Toaster />
